conv_3_3 with ubuffer and old extraction unit tests pass with expand_expr consolidated to HWUtils
