+++ src/HWUtils.h
@@ -5,6 +5,7 @@
 #include <vector>
 
 #include "IR.h"
+#include "Scope.h"
 
 namespace Halide {
   namespace Internal {
@@ -14,6 +15,10 @@ namespace Halide {
     std::ostream& operator<<(std::ostream& os, const std::vector<std::string>& vec);
     bool is_parallelized(const For *op);
     std::string first_for_name(Stmt s);
+    bool provide_at_level(Stmt s, std::string var);
+    bool call_at_level(Stmt s, std::string var);
+    Expr expand_expr(Expr e, const Scope<Expr> &scope);
+    Stmt expand_expr(Stmt e, const Scope<Expr> &scope);
   }
 }
 #endif
